Update flow.py
This is the first version of the gating. I still have to add the changes we talked about in lab meeting. Also, I want to break down the functions further. As it is now, each function gates either NK or T cells and generates a graph. I think it would be helpful to make graph, cell count, graphs, and data storage (etc.) (which are present in both NK and T fxns) into separate functions.